Color change functions outsourced
