+++ frontend/src/routes/+page.svelte
@@ -5,6 +5,7 @@
 	import ScoreForm from '$lib/components/ScoreForm.svelte';
 	import Timer from '$lib/components/Timer.svelte';
 	import { background } from '$lib/stores.js';
+	import { hexToRgb, hsvToRgb, rgbToHex, rgbToHsv } from '$lib/utils.js';
 
 
 	let state = 1;
@@ -64,31 +65,10 @@
 		}
 	});
 
-
-	const hexToRgb = hex =>
-		hex.replace(/^#?([a-f\d])([a-f\d])([a-f\d])$/i
-			,(m, r, g, b) => '#' + r + r + g + g + b + b)
-		.substring(1).match(/.{2}/g)
-		.map(x => parseInt(x, 16));
-
-	const rgbToHex = (r, g, b) => 
-		'#' + (1 << 24 | r << 16 | g << 8 | b).toString(16).slice(1);
-
-	const rgb2hsv = (r, g, b) => {
- 		let v=Math.max(r,g,b), c=v-Math.min(r,g,b);
-		let h= c && ((v==r) ? (g-b)/c : ((v==g) ? 2+(b-r)/c : 4+(r-g)/c)); 
-		return [60*(h<0?h+6:h), v&&c/v, v];
-	};
-
-	const hsv2rgb = (h, s, v) => {
-		let f= (n,k=(n+h/60)%6) => v - v*s*Math.max(Math.min(k,4-k,1), 0);
-		return [f(5),f(3),f(1)];
-	};
-
 	$: level, (() => {
 		let changeFactor = 30 * (1 + Math.floor(Math.random() * 10));
-		let [h,s,v] = rgb2hsv(...hexToRgb($background)); 
-		background.set(rgbToHex(...hsv2rgb((h + changeFactor) % 360, s, v)));
+		let [h,s,v] = rgbToHsv(...hexToRgb($background)); 
+		background.set(rgbToHex(...hsvToRgb((h + changeFactor) % 360, s, v)));
 	})();
 </script>
